# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Ticketing support microservice",
        "title": "Gin Swagger Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Cypodsolutions",
            "url": "http://www.cypod.com/",
            "email": "naders@cypodsolutions.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/payments_logs": {
            "get": {
                "description": "List payments, filtering by confirmation \u0026 customer id are optional params, pagination params is required",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payments logs"
                ],
                "summary": "List Payments",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page ID",
                        "name": "page_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page Size",
                        "name": "page_size",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "customer_id",
                        "name": "customer_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/db.PaymentLog"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/payments_logs/{log_id}": {
            "put": {
                "description": "Update Payment log's due date \u0026 confirmation using the payment log ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payments logs"
                ],
                "summary": "Update Payment",
                "parameters": [
                    {
                        "description": "Update log body",
                        "name": "arg",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.updatePaymentLogRequestJSON"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "payment log ID for update",
                        "name": "log_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/db.PaymentLog"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            },
            "delete": {
                "description": "Delete a payment log, removing its corresponding charges from the customer total charges",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "payments logs"
                ],
                "summary": "Delete Payment Log",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Log ID",
                        "name": "log_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "true"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        }
    },
    "definitions": {
...
        "api.updatePaymentLogRequestJSON": {
            "type": "object",
            "required": [
                "confirmed"
            ],
            "properties": {
                "confirmed": {
                    "type": "boolean"
                },
                "due_date": {
                    "$ref": "#/definitions/sql.NullTime"
                }
            }
        },
...
        "db.PaymentLog": {
            "type": "object",
            "properties": {
                "confirmation_date": {
                    "$ref": "#/definitions/sql.NullTime"
                },
                "confirmed": {
                    "type": "boolean"
                },
                "customer_id": {
                    "type": "integer"
                },
                "due_date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "order_id": {
                    "type": "integer"
                },
                "payment": {
                    "type": "number"
                }
            }
        },
...
        "sql.NullTime": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        }
    }
}